Refactorization

+++ lecture_extractor.py
@@ -1,13 +1,13 @@
 #!/bin/python3
 
-import math, operator
-from PIL import Image
-import sys
-import os
-import glob
-import subprocess
-import shutil
 import imagehash
+import glob
+import os
+import shutil
+import subprocess
+import sys
+
+from PIL import Image
 from tqdm import tqdm
 
 
@@ -30,42 +30,43 @@ class LectureExtractor(object):
         self.threshold = 5
 
 
-    def compare(self, image1=None, image2=None):
-        hash1 = imagehash.average_hash(image1)
-        hash2 = imagehash.average_hash(image2)
+    def deduplicate_frames(self, threshold=5):
+        if not os.path.exists("deduplicated"):
+            os.mkdir("deduplicated")
+            filelist = glob.glob(os.path.join("extraction", '*.png'))
+            filelist.sort()
+            count = 0
+            progress_bar = tqdm(range(0, len(filelist)))
+            progress_bar.set_description("Finding Slides")
+            for ii in progress_bar:
+                if ii < len(filelist)-1:
+                    image1 = Image.open(filelist[ii])
+                    image2 = Image.open(filelist[ii+1])
+                    hash1 = imagehash.average_hash(image1)
+                    hash2 = imagehash.average_hash(image2)
 
-        difference = hash1 - hash2
+                    difference = hash1 - hash2
 
-        return difference
+                    # **********************************************************
+                    # TODO: Need logic here to index the found slides, including
+                    # first and last appearances, as well as duration
+                    # **********************************************************
 
-
-    def deduplicate(self, location, threshold=5):
-        filelist = glob.glob(os.path.join("extraction", '*.png'))
-        filelist.sort()
-        count = 0
-        progress_bar = tqdm(range(0, len(filelist)))
-        progress_bar.set_description("Finding Slides:")
-        for ii in progress_bar:
-            if ii < len(filelist)-1:
-                image1 = Image.open(filelist[ii])
-                image2 = Image.open(filelist[ii+1])
-
-                difference = self.compare(image1, image2)
-                if difference > 5:
-                    head, tail = os.path.split(filelist[ii])
-                    shutil.copyfile(filelist[ii], location + os.path.sep + tail)
+                    if difference > 5:
+                        head, tail = os.path.split(filelist[ii])
+                        shutil.copyfile(filelist[ii], "extraction" + os.path.sep + tail)
+                        count += 1
+                else:
+                    shutil.copyfile(filelist[ii], "extraction" + os.path.sep + tail)
                     count += 1
-            else:
-                shutil.copyfile(filelist[ii], location + os.path.sep + tail)
-                count += 1
+            
+            progress_bar.close()
+            shutil.rmtree("extraction")
         
-        del(progress_bar)
-        shutil.rmtree("extraction")
-
-        return count
+            print("Found {number_of_slides} slide(s)".format(number_of_slides=count))
 
 
-    def main(self):
+    def extract_frames(self):
         if not os.path.exists("extraction"):
             os.mkdir("extraction")
             ffmpeg_cmd = ["ffmpeg",
@@ -74,13 +75,11 @@ class LectureExtractor(object):
                         "-vf", "fps={sampling_rate}".format(sampling_rate=self.sampling_rate),
                         "extraction/{input_filename}_%010d.png".format(input_filename=self.input_filename)]
             subprocess.call(ffmpeg_cmd)
-            
-        if not os.path.exists("deduplicated"):
-            os.mkdir("deduplicated")
-            location = "deduplicated"
-            number_of_slides = self.deduplicate(location=location, threshold=self.threshold)
 
-            print("Found {number_of_slides} slide(s)".format(number_of_slides=number_of_slides))
+
+    def main(self):
+        self.extract_frames()
+        self.deduplicate_frames()
 
 
 if __name__ == '__main__':
